backend/pyproject.toml
[project]
name = "certificado-verde-blockchain"
version = "0.1.0"
description = "O Certificado Verde Blockchain (CVB) é um sistema projetado para garantir a autenticidade e rastreabilidade de certificados relacionados ao extrativismo legal, utilizando a tecnologia blockchain. O sistema visa combater fraudes e aumentar a transparência na cadeia produtiva de recursos florestais, como madeira, castanhas e óleos vegetais."
authors = [
    {name = "JomarJunior",email = "jomarjunior@poli.ufrj.br"}
]
license = "MIT"
readme = "README.md"
requires-python = "<3.15,>=3.10"
dependencies = [
    "typing-extensions (>=4.15.0,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "miraveja-di (>=0.1.0,<0.2.0)",
    "miraveja-log (>=0.1.0,<0.2.0)",
    "miraveja-authentication (>=0.1.1,<0.2.0)",
    "web3 (>=7.14.0,<8.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [
    { include = "certificado_verde_blockchain", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["miraveja_log"]

[tool.pylint.master]
ignore = ["tests", "venv", ".venv", "build", "dist", "__pycache__", "alembic"]
jobs = 0
module-naming-style = "snake_case"
const-naming-style = "UPPER_CASE"
class-naming-style = "PascalCase"
function-naming-style = "snake_case"
method-naming-style = "snake_case"
attr-naming-style = "snake_case"
argument-naming-style = "snake_case"
variable-naming-style = "snake_case"
good-names = ["certificado_verde_blockchain", "c", "T", "cls"]
load-plugins = ["pylint_pydantic"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "R0917",  # too-many-positional-arguments
    "W0622",  # redefined-builtin
    "W0718",  # broad-exception-caught
    "R0801",  # duplicate-code
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/certificado_verde_blockchain",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
    "--cov-fail-under=80"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
omit = [
    "__init__.py",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
[dependency-groups]
dev = [
    "pytest (>=8.2.0,<9.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pylint (>=4.0.3,<5.0.0)",
    "mypy (>=1.19.0,<2.0.0)",
    "pre-commit (>=4.5.0,<5.0.0)",
    "pylint-pytest (>=1.1.8,<2.0.0)",
    "pylint-pydantic (>=0.4.1,<0.5.0)"
]
